NEW Controllers. Formato hora en 24HS
