+++ backend/controllers/authController.js
@@ -1,0 +1,53 @@
+const User = require('../models/User');
+const jwt = require('jsonwebtoken');
+const bcrypt = require('bcryptjs');
+
+exports.login = async (req, res) => {
+  const { email, password } = req.body;
+  try {
+    const user = await User.findOne({ email });
+    if (!user) return res.status(400).json({ message: 'Usuario no encontrado' });
+    const isMatch = await bcrypt.compare(password, user.password);
+    if (!isMatch) return res.status(400).json({ message: 'Contraseña incorrecta' });
+    const token = jwt.sign({ id: user._id, role: user.role }, process.env.JWT_SECRET, { expiresIn: '1h' });
+    res.json({ token, role: user.role });
+  } catch (err) {
+    res.status(500).json({ message: 'Error en el servidor' });
+  }
+};
+
+exports.register = async (req, res) => {
+  const { email, password, firstName, lastName, role } = req.body;
+  try {
+    const existingUser = await User.findOne({ email });
+    if (existingUser) {
+      return res.status(400).json({ message: 'El email ya está registrado' });
+    }
+    const newUser = new User({
+      email,
+      password,
+      firstName,
+      lastName,
+      role: role || 'employee'
+    });
+    await newUser.save();
+    const token = jwt.sign(
+      { id: newUser._id, role: newUser.role },
+      process.env.JWT_SECRET,
+      { expiresIn: '1h' }
+    );
+    res.status(201).json({ token, role: newUser.role });
+  } catch (err) {
+    res.status(500).json({ message: 'Error al registrar usuario' });
+  }
+};
+
+exports.getMe = async (req, res) => {
+  try {
+    const user = await User.findById(req.user.id).select('email firstName lastName');
+    if (!user) return res.status(404).json({ message: 'Usuario no encontrado' });
+    res.json(user);
+  } catch (err) {
+    res.status(500).json({ message: 'Error al obtener usuario' });
+  }
+};
