+++ backend/controllers/profileController.js
@@ -1,0 +1,14 @@
+const User = require('../models/User');
+exports.editProfile = async (req, res) => {
+  try {
+    const user = await User.findById(req.user.id);
+    if (!user) return res.status(404).json({ message: 'Usuario no encontrado' });
+    user.firstName = req.body.firstName || user.firstName;
+    user.lastName = req.body.lastName || user.lastName;
+    if (req.body.password) user.password = req.body.password;
+    await user.save();
+    res.json({ firstName: user.firstName, lastName: user.lastName });
+  } catch (err) {
+    res.status(500).json({ message: 'Error al actualizar perfil' });
+  }
+};
