+++ backend/routes/profile.js
@@ -1,26 +1,8 @@
 const express = require('express');
-const router = express.Router();
 const { auth } = require('../middleware/auth');
-const User = require('../models/User');
+const profileController = require('../controllers/profileController');
+const router = express.Router();
 
-// PUT /api/profile - Editar perfil propio (sin cambiar email)
-router.put('/', auth, async (req, res) => {
-  try {
-    const user = await User.findById(req.user.id);
-    if (!user) return res.status(404).json({ message: 'Usuario no encontrado' });
-
-    user.firstName = req.body.firstName || user.firstName;
-    user.lastName = req.body.lastName || user.lastName;
-    if (req.body.password) user.password = req.body.password;
-
-    await user.save();
-    res.json({
-      firstName: user.firstName,
-      lastName: user.lastName
-    });
-  } catch (err) {
-    res.status(500).json({ message: 'Error al actualizar perfil' });
-  }
-});
+router.put('/', auth, profileController.editProfile);
 
 module.exports = router;
