+++ backend/routes/subRole.js
@@ -1,66 +1,12 @@
 const express = require('express');
-const router = express.Router();
 const { auth, isAdmin } = require('../middleware/auth');
-const SubRole = require('../models/SubRole');
+const subRoleController = require('../controllers/subRoleController');
+const router = express.Router();
 
-// POST /api/subroles/ - Crear SubRole 
-router.post('/', auth, isAdmin, async (req, res) => {
-  try {
-    const { description, price, extraPrice } = req.body;
-    const subRole = new SubRole({ description, price, extraPrice });
-    await subRole.save();
-    res.status(201).json(subRole);
-  } catch (err) {
-    res.status(500).json({ message: 'Error al crear SubRole' });
-  }
-});
-
-// GET /api/subroles/ - Listar todos los SubRoles
-router.get('/', auth, isAdmin, async (req, res) => {
-  try {
-    const subRoles = await SubRole.find();
-    res.json(subRoles);
-  } catch (err) {
-    res.status(500).json({ message: 'Error al obtener SubRoles' });
-  }
-});
-
-// GET /api/subroles/:id - Obtener un SubRole por ID
-router.get('/:id', auth, isAdmin, async (req, res) => {
-  try {
-    const subRole = await SubRole.findById(req.params.id);
-    if (!subRole) return res.status(404).json({ message: 'SubRole no encontrado' });
-    res.json(subRole);
-  } catch (err) {
-    res.status(500).json({ message: 'Error al obtener SubRole' });
-  }
-});
-
-// PUT /api/subroles/:id - Actualizar un SubRole
-router.put('/:id', auth, isAdmin, async (req, res) => {
-  try {
-    const { description, price, extraPrice } = req.body;
-    const subRole = await SubRole.findByIdAndUpdate(
-      req.params.id,
-      { description, price, extraPrice },
-      { new: true }
-    );
-    if (!subRole) return res.status(404).json({ message: 'SubRole no encontrado' });
-    res.json(subRole);
-  } catch (err) {
-    res.status(500).json({ message: 'Error al actualizar SubRole' });
-  }
-});
-
-// DELETE /api/subroles/:id - Eliminar un SubRole
-router.delete('/:id', auth, isAdmin, async (req, res) => {
-  try {
-    const subRole = await SubRole.findByIdAndDelete(req.params.id);
-    if (!subRole) return res.status(404).json({ message: 'SubRole no encontrado' });
-    res.json({ message: 'SubRole eliminado' });
-  } catch (err) {
-    res.status(500).json({ message: 'Error al eliminar SubRole' });
-  }
-});
+router.post('/', auth, isAdmin, subRoleController.create);
+router.get('/', auth, isAdmin, subRoleController.getAll);
+router.get('/:id', auth, isAdmin, subRoleController.getOne);
+router.put('/:id', auth, isAdmin, subRoleController.update);
+router.delete('/:id', auth, isAdmin, subRoleController.delete);
 
 module.exports = router;
